ElementType->Element; rename trait, impl and module; remove missing license check
