--- primitives/src/merkle_tree/merkle_tree_frontier.rs
@@ -1,1 +1,0 @@
-// TODO(Chengyu): place holder for future implementation.
